backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
transformers>=4.28.1
torch>=2.0.0
numpy==1.23.5
spacy>=3.5.2
pypdf>=3.8.1
python-docx>=0.8.11
nltk>=3.8.1
scikit-learn>=1.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl